+++ cub/device/device_reduce.cuh
@@ -363,8 +363,8 @@ struct DeviceReduce
         cudaError error = cudaSuccess;
         do
         {
-            if (stream_synchronous) CubLog("Invoking ReduceSingle<<<1, %d, 0, %d>>>(), %d items per thread\n",
-                single_block_dispatch_params.block_threads, (int) stream, single_block_dispatch_params.items_per_thread);
+            if (stream_synchronous) CubLog("Invoking ReduceSingle<<<1, %d, 0, %lld>>>(), %d items per thread\n",
+                single_block_dispatch_params.block_threads, (long long) stream, single_block_dispatch_params.items_per_thread);
 
             // Invoke ReduceSingle
             single_block_kernel<<<1, single_block_dispatch_params.block_threads>>>(
@@ -494,7 +494,7 @@ struct DeviceReduce
                 #ifndef __CUDA_ARCH__
 
                     // We're on the host, so prepare queue on device (because its faster than if we prepare it here)
-                    if (stream_synchronous) CubLog("Invoking prepare_drain_kernel<<<1, 1, 0, %d>>>()\n", (int) stream);
+                    if (stream_synchronous) CubLog("Invoking prepare_drain_kernel<<<1, 1, 0, %lld>>>()\n", (long long) stream);
 
                     prepare_drain_kernel<<<1, 1, 0, stream>>>(queue, num_items);
 
@@ -515,8 +515,8 @@ struct DeviceReduce
             if (CubDebug(error = DeviceAllocate((void**) &d_block_partials, multi_grid_size * sizeof(T), device_allocator))) break;
 
             // Invoke MultiBlockReduce
-            if (stream_synchronous) CubLog("Invoking multi_block_kernel<<<%d, %d, 0, %d>>>(), %d items per thread, %d SM occupancy\n",
-                multi_grid_size, multi_block_dispatch_params.block_threads, (int) stream, multi_block_dispatch_params.items_per_thread, multi_sm_occupancy);
+            if (stream_synchronous) CubLog("Invoking multi_block_kernel<<<%d, %d, 0, %lld>>>(), %d items per thread, %d SM occupancy\n",
+                multi_grid_size, multi_block_dispatch_params.block_threads, (long long) stream, multi_block_dispatch_params.items_per_thread, multi_sm_occupancy);
 
             multi_block_kernel<<<multi_grid_size, multi_block_dispatch_params.block_threads, 0, stream>>>(
                 d_in,
@@ -535,8 +535,8 @@ struct DeviceReduce
             #endif
 
             // Invoke SingleBlockReduce
-            if (stream_synchronous) CubLog("Invoking single_block_kernel<<<%d, %d, 0, %d>>>(), %d items per thread\n",
-                1, single_block_dispatch_params.block_threads, (int) stream, single_block_dispatch_params.items_per_thread);
+            if (stream_synchronous) CubLog("Invoking single_block_kernel<<<%d, %d, 0, %lld>>>(), %d items per thread\n",
+                1, single_block_dispatch_params.block_threads, (long long) stream, single_block_dispatch_params.items_per_thread);
 
             single_block_kernel<<<1, single_block_dispatch_params.block_threads, 0, stream>>>(
                 d_block_partials,
